# Conflicts:
#	team4Project/src/main/java/com/web/store/account/controller/MemberController.java
#	team4Project/src/main/java/com/web/store/account/dao/impl/MemberDaoImpl.java
#	team4Project/src/main/java/com/web/store/account/service/impl/AccountServiceImpl.java
#	team4Project/src/main/java/com/web/store/home/controller/HomeController.java
#	team4Project/src/main/webapp/WEB-INF/views/layout/accountMenu.jsp

+++ team4Project/src/main/java/com/web/store/product/model/Product.java
@@ -39,7 +39,7 @@ public class Product implements Serializable{
 	private String productType;
 	@Column(name= "ProductDescript")
 	private String productDescript;
-//	@ManyToOne(fetch = FetchType.LAZY,cascade = CascadeType.ALL)    // javax.persistence.CascadeType;
+//	@ManyToOne(fetch = FetchType.LAZY,cascade = CascadeType.ALL)    
 //	@JoinColumn(name= "CompanyName")
 	private String companyName;
 	
@@ -48,7 +48,7 @@ public class Product implements Serializable{
 	@Column(name= "ProductPrice")
 	private Integer productPrice;
 	@Column(name= "Discount")
-	private Integer discount=1;
+	private Double discount= 1.0;
 	@Column(name= "ProductPic")
 	private Blob productPic;
 	@Column(name= "PicName")
@@ -66,7 +66,7 @@ public class Product implements Serializable{
 	
 	
 	public  Product(Integer productId,String productName,String productType,String productDescript,
-			String companyName,Integer productStuck,Integer productPrice,Integer discount,Blob productPic
+			String companyName,Integer productStuck,Integer productPrice,Double discount,Blob productPic
 			,Integer status){
 		super();
 		this.productId = productId;
@@ -81,7 +81,7 @@ public class Product implements Serializable{
 		this.status=status;
 	}
 	public  Product(String productName,String productType,String productDescript,
-			String companyName,Integer productStuck,Integer productPrice,Integer discount,Blob productPic,Integer status
+			String companyName,Integer productStuck,Integer productPrice,Double discount,Blob productPic,Integer status
 			){
 		super();
 		this.productName =productName;
@@ -172,10 +172,10 @@ public class Product implements Serializable{
 		this.productPrice = productPrice;
 	}
 	
-	public Integer getdiscount() {
+	public Double getdiscount() {
 		return discount;
 	}
-	public void setdiscount(Integer discount) {
+	public void setdiscount(Double discount) {
 		this.discount = discount;
 	}
 	public Blob getProductPic() {
